# Conflicts:
#	dcoj-web/src/main/java/com/dcoj/entity/ProblemEntity.java
#	dcoj-web/src/main/java/com/dcoj/service/ProblemService.java

+++ dcoj-web/src/main/java/com/dcoj/service/TagService.java
@@ -64,7 +64,7 @@ public interface TagService {
      * @param tid
      * @param flag
      */
-    void updateTagUsedTimes(long tid, boolean flag);
+    void updateTagUsedTimes(int tid, boolean flag);
 
 
     /**
